+++ airbyte-integrations/connectors/source-postgres/src/main/java/io/airbyte/integrations/source/postgres/DebeziumRecordPublisher.java
@@ -1,0 +1,162 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020 Airbyte
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package io.airbyte.integrations.source.postgres;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import io.airbyte.protocol.models.AirbyteStream;
+import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
+import io.airbyte.protocol.models.ConfiguredAirbyteStream;
+import io.debezium.engine.ChangeEvent;
+import io.debezium.engine.DebeziumEngine;
+import io.debezium.engine.format.Json;
+import io.debezium.engine.spi.OffsetCommitPolicy.AlwaysCommitOffsetPolicy;
+import java.util.Properties;
+import java.util.Queue;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.Collectors;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.codehaus.plexus.util.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class DebeziumRecordPublisher implements AutoCloseable {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(DebeziumRecordPublisher.class);
+  private final ExecutorService executor;
+  private DebeziumEngine<ChangeEvent<String, String>> engine;
+
+  private final JsonNode config;
+  private final ConfiguredAirbyteCatalog catalog;
+  private final AirbyteFileOffsetBackingStore offsetManager;
+
+  private final AtomicBoolean hasClosed;
+  private final AtomicReference<Throwable> thrownError;
+
+  public DebeziumRecordPublisher(JsonNode config, ConfiguredAirbyteCatalog catalog, AirbyteFileOffsetBackingStore offsetManager) {
+    this.config = config;
+    this.catalog = catalog;
+    this.offsetManager = offsetManager;
+    this.hasClosed = new AtomicBoolean(false);
+    this.thrownError = new AtomicReference<>();
+    this.executor = Executors.newSingleThreadExecutor();
+  }
+
+  public void start(Queue<ChangeEvent<String, String>> queue) {
+    engine = DebeziumEngine.create(Json.class)
+        .using(getDebeziumProperties(config, catalog, offsetManager))
+        .using(new AlwaysCommitOffsetPolicy())
+        .notifying(queue::add)
+        .using((success, message, error) -> {
+          LOGGER.info("completed!");
+          thrownError.set(error);
+        })
+        .build();
+
+    // Run the engine asynchronously ...
+    executor.execute(engine);
+  }
+
+  public boolean hasClosed() {
+    return hasClosed.get();
+  }
+
+  public synchronized void close() throws Exception {
+    if (!hasClosed.get()) {
+      // consumers should assume records can be produced until engine has closed.
+      if (engine != null) {
+        engine.close();
+      }
+
+      // announce closure only engine is off.
+      hasClosed.set(true);
+
+      executor.shutdown();
+
+      if (thrownError.get() != null) {
+        throw new RuntimeException(thrownError.get());
+      }
+    }
+  }
+
+  // todo: make this use catalog as well
+  // todo: make this use the state for the files as well
+  protected static Properties getDebeziumProperties(JsonNode config, ConfiguredAirbyteCatalog catalog, AirbyteFileOffsetBackingStore offsetManager) {
+    final Properties props = new Properties();
+    props.setProperty("name", "engine");
+    props.setProperty("plugin.name", "pgoutput");
+    props.setProperty("connector.class", "io.debezium.connector.postgresql.PostgresConnector");
+    props.setProperty("offset.storage", "org.apache.kafka.connect.storage.FileOffsetBackingStore");
+    props.setProperty("offset.storage.file.filename", offsetManager.getOffsetFilePath().toString());
+    props.setProperty("offset.flush.interval.ms", "1000"); // todo: make this longer
+
+    // https://debezium.io/documentation/reference/configuration/avro.html
+    props.setProperty("key.converter.schemas.enable", "false");
+    props.setProperty("value.converter.schemas.enable", "false");
+
+    // https://debezium.io/documentation/reference/configuration/event-flattening.html
+    props.setProperty("delete.handling.mode", "rewrite");
+    props.setProperty("drop.tombstones", "false");
+    props.setProperty("transforms.unwrap.type", "io.debezium.transforms.ExtractNewRecordState");
+
+    final String tableWhitelist = getTableWhitelist(catalog);
+    props.setProperty("table.include.list", tableWhitelist);
+    props.setProperty("database.include.list", config.get("database").asText());
+    props.setProperty("name", "orders-postgres-connector");
+    props.setProperty("include_schema_changes", "true");
+    props.setProperty("database.server.name", "orders"); // todo
+    props.setProperty("database.hostname", config.get("host").asText());
+    props.setProperty("database.port", config.get("port").asText());
+    props.setProperty("database.user", config.get("username").asText());
+
+    if (config.has("password")) {
+      props.setProperty("database.password", config.get("password").asText());
+    }
+
+    props.setProperty("database.dbname", config.get("database").asText());
+    props.setProperty("database.history", "io.debezium.relational.history.FileDatabaseHistory"); // todo: any reason not to use in memory version and
+    // reload from
+    props.setProperty("database.history.file.filename", "/tmp/debezium/dbhistory-" + RandomStringUtils.randomAlphabetic(5) + ".dat");
+
+    props.setProperty("slot.name", config.get("replication_slot").asText());
+
+    props.setProperty("snapshot.mode", "exported"); // can use never if we want to manage full refreshes ourselves
+
+    return props;
+  }
+
+  protected static String getTableWhitelist(ConfiguredAirbyteCatalog catalog) {
+    return catalog.getStreams().stream()
+        .map(ConfiguredAirbyteStream::getStream)
+        .map(AirbyteStream::getName)
+        // debezium needs commas escaped to split properly
+        .map(x -> StringUtils.escape(x, new char[] {','}, "\\,"))
+        .collect(Collectors.joining(","));
+
+  }
+
+}
