+++ airbyte-integrations/connectors/source-postgres/src/main/java/io/airbyte/integrations/source/postgres/PostgresSource.java
@@ -24,16 +24,11 @@
 
 package io.airbyte.integrations.source.postgres;
 
-import static java.lang.Thread.sleep;
-
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.ImmutableMap;
-import io.airbyte.commons.concurrency.CloseableLinkedBlockingQueue;
 import io.airbyte.commons.functional.CheckedConsumer;
 import io.airbyte.commons.json.Jsons;
-import io.airbyte.commons.lang.Queues;
 import io.airbyte.commons.util.AutoCloseableIterator;
 import io.airbyte.commons.util.AutoCloseableIterators;
 import io.airbyte.commons.util.CompositeIterator;
@@ -50,18 +45,13 @@ import io.airbyte.integrations.source.jdbc.JdbcStateManager;
 import io.airbyte.protocol.models.AirbyteCatalog;
 import io.airbyte.protocol.models.AirbyteMessage;
 import io.airbyte.protocol.models.AirbyteMessage.Type;
-import io.airbyte.protocol.models.AirbyteRecordMessage;
 import io.airbyte.protocol.models.AirbyteStateMessage;
 import io.airbyte.protocol.models.AirbyteStream;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.protocol.models.ConfiguredAirbyteStream;
 import io.airbyte.protocol.models.SyncMode;
 import io.debezium.engine.ChangeEvent;
-import io.debezium.engine.DebeziumEngine;
-import io.debezium.engine.format.Json;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.sql.SQLException;
@@ -71,19 +61,10 @@ import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Properties;
 import java.util.Set;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
-import org.apache.commons.lang3.RandomStringUtils;
-import org.apache.kafka.connect.data.Struct;
-import org.apache.kafka.connect.source.SourceRecord;
+import java.util.concurrent.LinkedBlockingQueue;
 import org.codehaus.plexus.util.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -167,76 +148,6 @@ public class PostgresSource extends AbstractJdbcSource implements Source {
     }
   }
 
-  enum SnapshotMetadata {
-    TRUE,
-    FALSE,
-    LAST
-  }
-
-  private PgLsn extractLsn(ChangeEvent<String, String> event) {
-    return Optional.ofNullable(event.value())
-        .flatMap(value -> Optional.ofNullable(Jsons.deserialize(value).get("source")))
-        .flatMap(source -> Optional.ofNullable(source.get("lsn").asText()))
-        .map(Long::parseLong)
-        .map(PgLsn::fromLong)
-        .orElseThrow(() -> new IllegalStateException("Could not find LSN"));
-  }
-
-  private SnapshotMetadata getSnapshotMetadata(ChangeEvent<String, String> event) {
-    try {
-      final Method sourceRecordMethod = event.getClass().getMethod("sourceRecord");
-      sourceRecordMethod.setAccessible(true);
-      final SourceRecord sourceRecord = (SourceRecord) sourceRecordMethod.invoke(event);
-      final String snapshot = ((Struct) sourceRecord.value()).getStruct("source").getString("snapshot");
-
-      if (snapshot == null) {
-        return null;
-      }
-
-      // the snapshot field is an enum of true, false, and last.
-      return SnapshotMetadata.valueOf(snapshot.toUpperCase());
-    } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private Predicate<ChangeEvent<String, String>> getTerminationPredicate(JdbcDatabase database, JdbcStateManager stateManager) {
-    final PgLsn targetLsn = getLsn(database);
-    LOGGER.info("identified target lsn: " + targetLsn);
-
-    return (event) -> {
-      final PgLsn eventLsn = extractLsn(event);
-
-      if (targetLsn.compareTo(eventLsn) > 0) {
-        return false;
-      } else {
-        final SnapshotMetadata snapshotMetadata = getSnapshotMetadata(event);
-        // if not snapshot or is snapshot but last record in snapshot.
-        return SnapshotMetadata.TRUE != snapshotMetadata;
-      }
-    };
-  }
-
-  private static class DebeziumPayload {
-
-    private final ChangeEvent<String, String> changeEvent;
-    private final AirbyteMessage message;
-
-    public DebeziumPayload(final ChangeEvent<String, String> changeEvent, final AirbyteMessage recordMessage) {
-      this.changeEvent = changeEvent;
-      this.message = recordMessage;
-    }
-
-    public ChangeEvent<String, String> getChangeEvent() {
-      return changeEvent;
-    }
-
-    public AirbyteMessage getMessage() {
-      return message;
-    }
-
-  }
-
   private AirbyteFileOffsetBackingStore initializeState(JdbcStateManager stateManager) {
     final Path cdcWorkingDir;
     try {
@@ -265,83 +176,26 @@ public class PostgresSource extends AbstractJdbcSource implements Source {
       // incremental updating of the state structs in the CDC impl.
       final AirbyteFileOffsetBackingStore offsetManager = initializeState(stateManager);
 
-      final Predicate<ChangeEvent<String, String>> hasReachedLsnPredicate = getTerminationPredicate(database, stateManager);
-      AtomicReference<Throwable> thrownError = new AtomicReference<>();
-      AtomicBoolean completed = new AtomicBoolean(false);
-      ExecutorService executor = Executors.newSingleThreadExecutor();
-      final CloseableLinkedBlockingQueue<DebeziumPayload> queue = new CloseableLinkedBlockingQueue<>(executor::shutdown);
+      final PgLsn targetLsn = getLsn(database);
+      LOGGER.info("identified target lsn: " + targetLsn);
 
-      DebeziumEngine<ChangeEvent<String, String>> engine = DebeziumEngine.create(Json.class)
-          .using(getDebeziumProperties(config, catalog, offsetManager))
-          .notifying(event -> {
-            try {
-              final AirbyteMessage message = convertChangeEvent(event, emittedAt);
-              queue.add(new DebeziumPayload(event, message));
-            } catch (Exception e) {
-              LOGGER.info("error: " + e);
-              thrownError.set(e);
-            }
-          })
-          .using((success, message, error) -> {
-            LOGGER.info("completed!");
-            completed.set(true);
-            thrownError.set(error);
-          })
-          .build();
+      final LinkedBlockingQueue<ChangeEvent<String, String>> queue = new LinkedBlockingQueue<>();
 
-      // Run the engine asynchronously ...
-      executor.execute(engine);
+      final DebeziumRecordPublisher publisher = new DebeziumRecordPublisher(config, catalog, offsetManager);
+      publisher.start(queue);
 
-      final Iterator<DebeziumPayload> queueIterator = Queues.toStream(queue).iterator();
-      final AbstractIterator<DebeziumPayload> iterator = new AbstractIterator<>() {
+      // handle state machine around pub/sub logic.
+      final AutoCloseableIterator<ChangeEvent<String, String>> eventIterator = new DebeziumRecordConsumer(
+          queue,
+          targetLsn,
+          publisher::hasClosed,
+          publisher::close,
+          emittedAt);
 
-        private boolean hasReachedLsn = false;
-        private int sleepCount = 0;
-
-        @Override
-        protected DebeziumPayload computeNext() {
-          // if we have reached the lsn we stop, otherwise we have the potential to wait indefinitely for the
-          // next value.
-          if (!hasReachedLsn) {
-            while (!queueIterator.hasNext()) {
-              if (sleepCount > 5) {
-                return endOfData();
-              }
-              LOGGER.info("sleeping. count: {}", sleepCount++);
-              try {
-                sleep(5000);
-              } catch (InterruptedException e) {
-                throw new RuntimeException(e);
-              }
-            }
-
-            final DebeziumPayload next = queueIterator.next();
-            LOGGER.info("next {}", next);
-            // todo fix this cast. the record passed to this iterator has to include the lsn somewhere. it can
-            // either be the full change event or some smaller object that just includes the lsn.
-            // we guarantee that this will always eventually return true, because we pick an LSN that already
-            // exists when we start the sync.
-            if (hasReachedLsnPredicate.test(next.getChangeEvent())) {
-              hasReachedLsn = true;
-            }
-            return next;
-          } else {
-            return endOfData();
-          }
-        }
-
-      };
-
-      final AutoCloseableIterator<DebeziumPayload> payloadIterator = AutoCloseableIterators.fromIterator(iterator, () -> {
-        engine.close();
-        executor.shutdown();
-
-        if (thrownError.get() != null) {
-          throw new RuntimeException(thrownError.get());
-        }
-      });
-
-      final AutoCloseableIterator<AirbyteMessage> messageIterator = AutoCloseableIterators.transform(payloadIterator, DebeziumPayload::getMessage);
+      // convert to airbyte message.
+      final AutoCloseableIterator<AirbyteMessage> messageIterator = AutoCloseableIterators.transform(
+          eventIterator,
+          (event) -> DebeziumEventUtils.convertChangeEvent(event, emittedAt));
 
       // our goal is to get the state at the time this supplier is called (i.e. after all message records
       // have been produced)
@@ -364,54 +218,6 @@ public class PostgresSource extends AbstractJdbcSource implements Source {
     } else {
       return super.getIncrementalIterators(config, database, catalog, tableNameToTable, stateManager, emittedAt);
     }
-  }
-
-  // todo: make this use catalog as well
-  // todo: make this use the state for the files as well
-  protected static Properties getDebeziumProperties(JsonNode config,
-                                                    ConfiguredAirbyteCatalog catalog,
-                                                    AirbyteFileOffsetBackingStore airbyteFileOffsetBackingStore) {
-    final Properties props = new Properties();
-    props.setProperty("name", "engine");
-    props.setProperty("plugin.name", "pgoutput");
-    props.setProperty("connector.class", "io.debezium.connector.postgresql.PostgresConnector");
-    props.setProperty("offset.storage", "org.apache.kafka.connect.storage.FileOffsetBackingStore");
-    props.setProperty("offset.storage.file.filename", airbyteFileOffsetBackingStore.getOffsetFilePath().toString());
-    props.setProperty("offset.flush.interval.ms", "1000"); // todo: make this longer
-
-    // https://debezium.io/documentation/reference/configuration/avro.html
-    props.setProperty("key.converter.schemas.enable", "false");
-    props.setProperty("value.converter.schemas.enable", "false");
-
-    // https://debezium.io/documentation/reference/configuration/event-flattening.html
-    props.setProperty("delete.handling.mode", "rewrite");
-    props.setProperty("drop.tombstones", "false");
-    props.setProperty("transforms.unwrap.type", "io.debezium.transforms.ExtractNewRecordState");
-
-    final String tableWhitelist = getTableWhitelist(catalog);
-    props.setProperty("table.include.list", tableWhitelist);
-    props.setProperty("database.include.list", config.get("database").asText());
-    props.setProperty("name", "orders-postgres-connector");
-    props.setProperty("include_schema_changes", "true");
-    props.setProperty("database.server.name", "orders"); // todo
-    props.setProperty("database.hostname", config.get("host").asText());
-    props.setProperty("database.port", config.get("port").asText());
-    props.setProperty("database.user", config.get("username").asText());
-
-    if (config.has("password")) {
-      props.setProperty("database.password", config.get("password").asText());
-    }
-
-    props.setProperty("database.dbname", config.get("database").asText());
-    props.setProperty("database.history", "io.debezium.relational.history.FileDatabaseHistory"); // todo: any reason not to use in memory version and
-    // reload from
-    props.setProperty("database.history.file.filename", "/tmp/debezium/dbhistory-" + RandomStringUtils.randomAlphabetic(5) + ".dat");
-
-    props.setProperty("slot.name", config.get("replication_slot").asText());
-
-    props.setProperty("snapshot.mode", "exported"); // can use never if we want to manage full refreshes ourselves
-
-    return props;
   }
 
   protected static String getTableWhitelist(ConfiguredAirbyteCatalog catalog) {
@@ -453,44 +259,6 @@ public class PostgresSource extends AbstractJdbcSource implements Source {
     properties.set("_ab_cdc_deleted_at", numberType);
 
     return stream;
-  }
-
-  public static AirbyteMessage convertChangeEvent(ChangeEvent<String, String> event, Instant emittedAt) {
-    final JsonNode debeziumRecord = Jsons.deserialize(event.value());
-    final JsonNode before = debeziumRecord.get("before");
-    final JsonNode after = debeziumRecord.get("after");
-    final JsonNode source = debeziumRecord.get("source");
-
-    final JsonNode data = formatDebeziumData(before, after, source);
-
-    final String streamName = source.get("table").asText();
-
-    final AirbyteRecordMessage airbyteRecordMessage = new AirbyteRecordMessage()
-        .withStream(streamName)
-        .withEmittedAt(emittedAt.toEpochMilli())
-        .withData(data);
-
-    return new AirbyteMessage()
-        .withType(AirbyteMessage.Type.RECORD)
-        .withRecord(airbyteRecordMessage);
-  }
-
-  public static JsonNode formatDebeziumData(JsonNode before, JsonNode after, JsonNode source) {
-    final ObjectNode base = (ObjectNode) (after.isNull() ? before : after);
-
-    long transactionMillis = source.get("ts_ms").asLong();
-    long lsn = source.get("lsn").asLong();
-
-    base.put("_ab_cdc_updated_at", transactionMillis);
-    base.put("_ab_cdc_lsn", lsn);
-
-    if (after.isNull()) {
-      base.put("_ab_cdc_deleted_at", transactionMillis);
-    } else {
-      base.put("_ab_cdc_deleted_at", (Long) null);
-    }
-
-    return base;
   }
 
   public static void main(String[] args) throws Exception {
